UX Improvements and Refactor Import/Export Styling
