+++ src/content/Settings/ExportImport/ExportImportPopups/ExportCalendarPopup.tsx
@@ -1,4 +1,4 @@
-import './ExportCalendarPopup.css'
+import './ExportImportPopup.css'
 import CloseIcon from '../../../Icons/CloseIcon'
 import { ISectionData } from '../../../App/App.types';
 import { useState } from 'react';
@@ -12,21 +12,21 @@ interface IProps {
 const ExportCalendarPopup = ({onCancel, sections, exportFunction}:IProps) => {
     const [selectedWorklist, setSelectedWorklist] = useState<number>(0)
     return (
-        <div className='ExportCalendarBackground'>
-            <div className='ExportCalendarPopup'>
-                <div className='ExportCalendarPopupHeader'>
+        <div className='CalendarBackground'>
+            <div className='CalendarPopup'>
+                <div className='CalendarPopupHeader'>
                     <div>Export Worklist</div>
-                    <div className='ExportCalendarPopupCloseButton' onClick={(onCancel)}>
+                    <div className='CalendarPopupCloseButton' onClick={(onCancel)}>
                         <CloseIcon size={16} color='white'/>
                     </div>
                 </div>
-                <div className='ExportCalendarPopupBody'>
+                <div className='CalendarPopupBody'>
                     Choose the worklist to export from: 
-                    <div className='ExportCalendarPopupButtonContainer'>
+                    <div className='CalendarPopupButtonContainer'>
                         <div>Worklist: </div>
                         {[0, 1, 2, 3].map((worklist) => 
                             <div 
-                                className='ExportCalendarPopupButton' 
+                                className='CalendarPopupButton' 
                                 style={selectedWorklist === worklist? {backgroundColor: "rgb(156, 232, 255)"}: {}}
                                 onClick={() => setSelectedWorklist(worklist)}
                             >
@@ -35,9 +35,9 @@ const ExportCalendarPopup = ({onCancel, sections, exportFunction}:IProps) => {
                             )
                         }
                     </div>
-                    <div className='ExportCalendarButtonContainer'>
-                    <div className='ExportCalendarButtonCancel' onClick={onCancel}>Cancel</div>
-                    <div className='ExportCalendarButtonConfirm' onClick={() => exportFunction(sections, selectedWorklist)}>Confirm</div>
+                    <div className='CalendarButtonContainer'>
+                    <div className='CalendarButtonCancel' onClick={onCancel}>Cancel</div>
+                    <div className='CalendarButtonConfirm' onClick={() => exportFunction(sections, selectedWorklist)}>Confirm</div>
                 </div>
                 </div>
             </div>
