--- src/content/Settings/ExportImport/ImportCalendarPopup/ImportCalendarPopup.css
@@ -1,110 +1,0 @@
-.ImportCalendarBackground {
-    background-color: rgba(99, 99, 99, 0.6);
-    position: fixed;
-    top: 0;
-    height: 650px;
-    width: 300px;
-    justify-content: center;
-    align-items: center;
-    display: flex;
-    z-index: 100;
-}
-
-.ImportCalendarPopup {
-    background-color: white;
-    width: 90%;
-    height: 30%;
-    border-radius: 10px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.ImportCalendarPopupHeader {
-    height: 15%;
-    background-color: #00A7E1;
-    color: white;
-    font-size: 15px;
-    border-top-left-radius: 10px;
-    border-top-right-radius: 10px;
-    font-weight: bold;
-    text-align: left;
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    padding-left: 10px;
-    padding-right: 10px;
-}
-
-.ImportCalendarPopupCloseButton {
-    cursor: pointer;
-    margin-right: 5px;
-}
-
-.ImportCalendarPopupBody {
-    height: 80%;
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.ImportCalendarPopupButton {
-    border-radius: 5px;
-    padding: 3px 5px;
-    margin-left: 3px;
-    cursor: pointer;
-    background-color: rgb(242, 252, 255);
-}
-
-.ImportCalendarPopupButton:hover {
-    background-color: rgb(156, 232, 255)
-}
-
-.ImportCalendarPopupButtonContainer {
-    display: flex;
-    align-items: center;
-}
-
-.ImportCalendarButtonConfirm {
-    width: 50%;
-    padding: 7px 20px;
-    font-size: 12px;
-    color: #fff;
-    background-color: #007bff;;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s, box-shadow 0.3s;
-    text-align: center;
-    margin: 10px 10px 2px 10px;
-}
-
-.ImportCalendarButtonConfirm:hover {
-    background-color: #0068d6;;
-}
-
-.ImportCalendarButtonContainer {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-}
-
-.ImportCalendarButtonCancel {
-    width: 50%;
-    padding: 7px 20px;
-    font-size: 12px;
-    color: #626262;
-    border: #626262;
-    border-width: 1px;
-    border-style: solid;
-    border-radius: 4px;
-    cursor: pointer;
-    text-align: center;
-    transition: background-color 0.3s, box-shadow 0.3s;
-    margin: 10px 10px 2px 10px;
-}
-
-.ImportCalendarButtonCancel:hover {
-    background-color: #d8d8d8;
-}
